+++ cli/root.go
@@ -33,6 +33,7 @@ func NewRootCommand(opts RootCommandOptions) *cobra.Command {
 	root.AddCommand(NewBuildCommand(BuildCommandOptions{IO: ioCfg, DefaultCatalogPath: opts.DefaultCatalogPath}))
 	root.AddCommand(NewValidateCommand(ValidateCommandOptions{IO: ioCfg, DefaultCatalogPath: opts.DefaultCatalogPath}))
 	root.AddCommand(NewModelsCommand(ModelsCommandOptions{IO: ioCfg, LoadBaseCatalog: opts.LoadBaseCatalog}))
+	root.AddCommand(NewSkillCommand(SkillCommandOptions{IO: ioCfg}))
 	root.InitDefaultCompletionCmd()
 	return root
 }